Split gce into its own package
Should insulate aws users from gce import
issues and make things more modular and nice

Testing: encrypted, updated and launched gce
images
