+++ brkt_cli/gce/encrypt_gce_image_args.py
