+++ brkt_cli/gce/launch_gce_image_args.py
