+++ brkt_cli/gce/update_encrypted_gce_image_args.py
